Changes made after review 3
